refactor: 입/출금 내역 퍼블리싱 수정
